Update `geo_rev` directly instead of using triggers
* Remove triggers, verify that tests fail
* Add explicit `geo_rev` updates, verify that tests pass
* Update places where instance was explicitly fetched to get updated geo_rev

Update `geo_rev` when:
* a map feature is created or updated - `treemap.views.map_feature`
* a map feature is deleted - `MapFeature.delete_with_user`
* random trees are created
